+++ test/unit/specs/elements/GeoSegmentedControl/GeoSegmentedControlItem.spec.js
@@ -4,7 +4,7 @@ import GeoInfoSegmentedControlItem from '@/elements/GeoSegmentedControl/GeoInfoS
 import GeoSuccessSegmentedControlItem from '@/elements/GeoSegmentedControl/GeoSuccessSegmentedControlItem.vue'
 import GeoWarningSegmentedControlItem from '@/elements/GeoSegmentedControl/GeoWarningSegmentedControlItem.vue'
 import GeoDangerSegmentedControlItem from '@/elements/GeoSegmentedControl/GeoDangerSegmentedControlItem.vue'
-import GeoTrimmedContent from '@/elements/GeoTrimmedContent/GeoTrimmedContent.vue'
+import GeoTrimmedContent from '@/elements/GeoTrimmedContent/GeoTrimmedContent'
 
 // create an extended `Vue` constructor
 const localVue = createLocalVue()
@@ -17,7 +17,7 @@ localVue.component('geo-trimmed-content', GeoTrimmedContent)
 
 describe('GeoSegmentedControlItem', () => {
   it('Should render button\'s content', function () {
-    const wrapper = getShallowWrapper({
+    const wrapper = shallowMount(GeoSegmentedControlItem, {
       slots: {
         default: '<span>Segmented control item title</span>'
       },
@@ -30,37 +30,37 @@ describe('GeoSegmentedControlItem', () => {
     expect(button.find('span').exists()).toBe(true)
   })
 
-  it('Should emit an event on click', function () {
-    const clickListener = jest.fn()
-    const wrapper = getShallowWrapper({
+  it('Should emit an event on click', function (done) {
+    const wrapper = shallowMount(GeoSegmentedControlItem, {
       propsData: {
         type: 'primary'
-      },
-      listeners: {
-        click: clickListener
       }
     })
     wrapper.find('.geo-segmented-control-item').trigger('click')
-    expect(clickListener).toHaveBeenCalled()
+    setTimeout(function () {
+      try {
+        expect(wrapper.emitted().click).toBeTruthy()
+        done()
+      } catch (error) {
+        done(error)
+      }
+    })
   })
 
   it('Should not emit an event when it\'s disabled', function () {
-    const clickListener = jest.fn()
-    const wrapper = getShallowWrapper({
+    const wrapper = shallowMount(GeoSegmentedControlItem, {
       propsData: {
         disabled: true
-      },
-      listeners: {
-        click: clickListener
       }
     })
 
-    wrapper.find('.geo-segmented-control-item').trigger('click')
-    expect(clickListener).not.toHaveBeenCalled()
+    const button = wrapper.find('.geo-segmented-control-item')
+    button.trigger('click')
+    expect(wrapper.emitted().click).toBeFalsy()
   })
 
   it('Should add disabled class when it\'s disabled', function () {
-    const wrapper = getShallowWrapper({
+    const wrapper = shallowMount(GeoSegmentedControlItem, {
       propsData: {
         disabled: true
       }
@@ -69,12 +69,14 @@ describe('GeoSegmentedControlItem', () => {
     expect(wrapper.find('.geo-segmented-control-item--disabled').exists()).toBe(true)
   })
 
-  it('Should not add disabled class when it\'s not disabled', function () {
-    expect(getShallowWrapper().find('.geo-segmented-control-item--disabled').exists()).toBe(false)
+  it('Should nont add disabled class when it\'s not disabled', function () {
+    const wrapper = shallowMount(GeoSegmentedControlItem)
+
+    expect(wrapper.find('.geo-segmented-control-item--disabled').exists()).toBe(false)
   })
 
   it('Should add active class when it\'s active', function () {
-    const wrapper = getShallowWrapper({
+    const wrapper = shallowMount(GeoSegmentedControlItem, {
       propsData: {
         active: true
       }
@@ -84,11 +86,13 @@ describe('GeoSegmentedControlItem', () => {
   })
 
   it('Should not add active class when it\'s not active', function () {
-    expect(getShallowWrapper().find('.geo-segmented-control-item--active').exists()).toBe(false)
+    const wrapper = shallowMount(GeoSegmentedControlItem)
+
+    expect(wrapper.find('.geo-segmented-control-item--active').exists()).toBe(false)
   })
 
   it('Should add outline class when it\'s outlined', function () {
-    const wrapper = getShallowWrapper({
+    const wrapper = shallowMount(GeoSegmentedControlItem, {
       propsData: {
         outline: true
       }
@@ -98,12 +102,10 @@ describe('GeoSegmentedControlItem', () => {
   })
 
   it('Should not add outline class when it\'s outlined', function () {
-    expect(getShallowWrapper().find('.geo-segmented-control-item--outline').exists()).toBe(false)
-  })
+    const wrapper = shallowMount(GeoSegmentedControlItem)
 
-  function getShallowWrapper (options) {
-    return shallowMount(GeoSegmentedControlItem, options)
-  }
+    expect(wrapper.find('.geo-segmented-control-item--outline').exists()).toBe(false)
+  })
 })
 
 const taxonomySegmentedControlItems = [
@@ -117,7 +119,7 @@ describe('GeoSegmentedControlItem Children', () => {
   taxonomySegmentedControlItems.forEach((taxonomySegmentedControlItem) => {
     describe(taxonomySegmentedControlItem.name, function () {
       it('Should render button\'s content', function () {
-        const wrapper = getShallowWrapper({
+        const wrapper = shallowMount(taxonomySegmentedControlItem, {
           slots: {
             default: '<span>Segmented control item title</span>'
           },
@@ -130,43 +132,46 @@ describe('GeoSegmentedControlItem Children', () => {
         expect(button.find('span').exists()).toBe(true)
       })
 
-      it('Should emit an event on click', function () {
-        const clickListener = jest.fn()
-        const wrapper = getShallowWrapper({
+      it('Should emit an event on click', function (done) {
+        const wrapper = shallowMount(taxonomySegmentedControlItem, {
           stubs: {
             GeoSegmentedControlItem
-          },
-          listeners: {
-            click: clickListener
           }
         })
-
         wrapper.find('.geo-segmented-control-item').trigger('click')
-        expect(clickListener).toHaveBeenCalled()
+        setTimeout(function () {
+          try {
+            expect(wrapper.emitted().click).toBeTruthy()
+            done()
+          } catch (error) {
+            done(error)
+          }
+        })
       })
 
-      it('Should not emit the click event when it\'s disabled', function () {
-        const clickListener = jest.fn()
-        const wrapper = getShallowWrapper({
-          propsData: {
-            disabled: true
+      it('Should not emit an event when it\'s disabled', function () {
+        const wrapper = shallowMount(taxonomySegmentedControlItem, {
+          context: {
+            props: {
+              disabled: true
+            }
           },
           stubs: {
             GeoSegmentedControlItem
-          },
-          listeners: {
-            click: clickListener
           }
         })
 
-        wrapper.find('.geo-segmented-control-item').trigger('click')
-        expect(clickListener).not.toHaveBeenCalled()
+        const button = wrapper.find('.geo-segmented-control-item')
+        button.trigger('click')
+        expect(wrapper.emitted().click).toBeFalsy()
       })
 
       it('Should add disabled class when it\'s disabled', function () {
-        const wrapper = getShallowWrapper({
-          propsData: {
-            disabled: true
+        const wrapper = shallowMount(taxonomySegmentedControlItem, {
+          context: {
+            props: {
+              disabled: true
+            }
           },
           stubs: {
             GeoSegmentedControlItem
@@ -177,7 +182,7 @@ describe('GeoSegmentedControlItem Children', () => {
       })
 
       it('Should not add disabled class when it\'s not disabled', function () {
-        const wrapper = getShallowWrapper({
+        const wrapper = shallowMount(taxonomySegmentedControlItem, {
           stubs: {
             GeoSegmentedControlItem
           }
@@ -187,9 +192,11 @@ describe('GeoSegmentedControlItem Children', () => {
       })
 
       it('Should add outline class when it\'s outlined', function () {
-        const wrapper = getShallowWrapper({
-          propsData: {
-            outline: true
+        const wrapper = shallowMount(taxonomySegmentedControlItem, {
+          context: {
+            props: {
+              outline: true
+            }
           },
           stubs: {
             GeoSegmentedControlItem
@@ -200,7 +207,7 @@ describe('GeoSegmentedControlItem Children', () => {
       })
 
       it('Should not add outline class when it\'s not outline', function () {
-        const wrapper = getShallowWrapper({
+        const wrapper = shallowMount(taxonomySegmentedControlItem, {
           stubs: {
             GeoSegmentedControlItem
           }
@@ -208,10 +215,6 @@ describe('GeoSegmentedControlItem Children', () => {
 
         expect(wrapper.find('.geo-segmented-control-item--outline').exists()).toBe(false)
       })
-
-      function getShallowWrapper (options) {
-        return shallowMount(taxonomySegmentedControlItem, options)
-      }
     })
   })
 })